moved removeTaintNodesBackingMachineSet unit tests to deployment_util_test
